Chunking image dataset is Done in Extract_features.

+++ dataloading/dataset.py
@@ -2,17 +2,17 @@ import os
 import tarfile
 import shutil
 import zipfile
-def prepare(dataset_path, dataset_size, batch_size, archive_files, verbose):
+def prepare(dataset_folder, dataset_size, batch_size, archive_files, chunk, verbose):
 
     num_batch = dataset_size // batch_size + (dataset_size % batch_size != 0)
     
+    dataset_path = f"{dataset_folder}/images"
     if not os.path.exists(dataset_path):
         os.makedirs(dataset_path)
-
-        
         if verbose:
                 print(" > Archive files:\n", archive_files)
-        temp = dataset_path +"/tmp"
+
+        temp = dataset_folder +"/tmp"
         # Unzip
         for archive in archive_files:
             os.makedirs(temp)
@@ -36,50 +36,62 @@ def prepare(dataset_path, dataset_size, batch_size, archive_files, verbose):
             shutil.rmtree(temp)
 
         print(" > Extracting dataset file complete!")
-    else:
+    if chunk:
         # remove any .txt files
-        files = os.listdir(dataset_path)
+        files = os.listdir(dataset_folder)
         for file in files:
             if file.endswith(".txt"):
-                file_path = os.path.join(dataset_path, file)
+                file_path = os.path.join(dataset_folder, file)
                 os.remove(file_path)
         if verbose:
                 print(f" > Removing .txt file complete!")
     
-    img_idx = 0
-    files = os.listdir(dataset_path)
-    files.sort()
-    for batch_index in range(num_batch):
-        file_path = f"{dataset_path}/batch{batch_index}.txt"
-        with open(file_path, 'w') as f:
-            while img_idx < (batch_index + 1 ) * batch_size:
-                # img_path = os.path.join(dataset_path, files[img_idx])
-                f.write(files[img_idx] + '\n')
-                img_idx += 1
-            f.close()
-        if verbose:
-                print(f" > Writing {file_path} complete!")
+        img_idx = 0
+        files = os.listdir(dataset_path)
+        files.sort()
+        for batch_index in range(num_batch):
+            file_path = f"{dataset_folder}/batch{batch_index}.txt"
+            with open(file_path, 'w') as f:
+                while img_idx < (batch_index + 1 ) * batch_size:
+                    # img_path = os.path.join(dataset_path, files[img_idx])
+                    f.write(files[img_idx] + '\n')
+                    img_idx += 1
+                f.close()
+            if verbose:
+                    print(f" > Writing {file_path} complete!")
     
     return num_batch
     
 def prepare_dataset(dataset_name, batch_size, 
                     source_dir="data",
+                    chunk=True,
                     verbose = False):
     archive_files = []
     dataset_size = 0
     if dataset_name == "holidays":
-        dataset_path = "results/holidays"
+        print(f" > Preparing {dataset_name} dataset...")
+        dataset_folder = "results/holidays"
         dataset_size = 1419
         archive_files = [f"{source_dir}\holidays\jpg{i}.tar.gz" for i in range(1,3)]
     elif dataset_name == "mirflickr1m":
-        dataset_path = "results/mirflickr1m"
+        print(f" > Preparing {dataset_name} dataset...")
+        dataset_folder = "results/mirflickr1m"
         dataset_size = 1000000
         archive_files = [f"{source_dir}\mirflickr1m\images{i}.zip" for i in range(8)]
+    elif dataset_name == "selected":
+        print(f" > Preparing {dataset_name} dataset...")
+        dataset_folder = "results/selected"
+        dataset_size = 50
+        archive_files = [f"{source_dir}\selected\selected.zip"]
+    else:
+         print(f" > The {dataset_name} dataset is not found!")
+         
 
-    num_batch = prepare(dataset_path=dataset_path,
+    num_batch = prepare(dataset_folder=dataset_folder,
                         dataset_size=dataset_size,
                         batch_size=batch_size,
                         archive_files=archive_files,
+                        chunk=chunk,
                         verbose=verbose
                         )
     print(f" > Preparing {dataset_name} dataset with batch_size {batch_size} is Done!")
